Refactored TempProject into its own module and started on tests
