+++ tests/integration/main.rs
@@ -1,6 +1,8 @@
 #[macro_use]
 mod util;
+mod temp_project;
 
 mod install;
 mod publish;
 mod read_projects;
+mod update;
